requirements.txt
pandas==2.0.3
numpy==1.26.4
requests==2.31.0
faiss-cpu==1.8.0
matplotlib==3.7.1
streamlit==1.34.0
streamlit-cropper==0.3.1
torch==2.1.2
